Struct andiskaz::color::CmyColor

pub struct CmyColor { /* fields omitted */ }

A CMY (Cyan-Magenta-Yellow) color. The lower one of its component is, the more it subtracts.

Implementations

impl CmyColor

pub const BASE: u8

Base of CMY colors (6).

pub fn try_new(cyan: u8, magenta: u8, yellow: u8) -> Result<Self, BadCmyColor>

Creates a new CmyColor given its components. Returns an error if any of the components is >= Self::BASE.

pub fn new(cyan: u8, magenta: u8, yellow: u8) -> Self

Creates a new CmyColor given its components.

Panics

Panics if any of the components is >= Self::BASE.

pub const fn cyan(self) -> u8

The level of cyan component.

pub const fn magenta(self) -> u8

The level of magenta component.

pub const fn yellow(self) -> u8

The level of yellow component.

pub const fn code(self) -> u8

The resulting code of the color.

pub fn set_cyan(self, cyan: u8) -> Self

Sets the cyan component.

Panics

Panics if the component is >= Self::BASE.

pub fn set_magenta(self, magenta: u8) -> Self

Sets the magenta component.

Panics

Panics if the component is >= Self::BASE.

pub fn set_yellow(self, yellow: u8) -> Self

Sets the yellow component.

Panics

Panics if the component is >= Self::BASE.