Module andiskaz::color::transform

This module defines color transformers. A color Transformer is simply a function over Colors. In fact, if you want to use a plain function as a Transformer, use FromFn. Analogously, there is a transformer over Color2s, the PairTransformer.

Structs

Adapt
Adapts a single color to match the given brightness.

AdaptBgToFg
Adapts background to match foreground.

AdaptFgToBg
Adapts foreground to match background.

Contrast
Adapts a single color to contrast the given brightness.

ContrastBgWithFg
Adapats background to contrast foreground.

ContrastFgWithBg
Adapats foreground to contrast background.

FromFn
Creates a transformer from a given function if the function has a signature similar to a Transformer or PairTransformer.

Id
Identity transformer. Preserves the input Color or Color2.

Invert
Inverts a single color.

PairWith
Updates a Color2 using the given foreground and background Transformers, each over a single color, by matching output foreground with color produced by foreground transformer, and output background with color produced by background transformer.

Seq
Sequence transformer. Applies the two given transformers in sequence on a color. Similar to function composition, but Seq(t, u) is analogous to u(t(color)) instead of t(u(color)) (the last one is how composition usually works).

SwapPairUsing
Updates a Color2 using the given foreground and background Transformers, each over a single color, like PairWith, but swapping foreground and background. This means input foreground is fed to a transformer which produces background, and input background is fed to a transformer which produces foreground.

Traits

PairTransformer
A color pair transformer. Changes a color pair.

Transformer
A color transformer. Changes a color.

Type Definitions

SwapPair
Just swaps foreground with background and nothing more.