Type Alias andiskaz::coord::Vec2

pub type Vec2 = Vec2<Coord>;

A coordinate made of two components x and y. The x axis corresponds to its expected meaning. When the feature plane is enabled, this is simply an alias to ::gardiz::coord::Vec2, otherwise it is a dedicated struct.

Aliased Type

struct Vec2 {
    pub y: u16,
    pub x: u16,
}

Fields
y: u16

Value of the Y-coordinate.

x: u16

Value of the X-coordinate.