Module andiskaz::error

This module exports error types used by the terminal handles.

Structs

Error
An error that may happen when executing an operation on the terminal.

EventsOff
Error returned by the events handler when the listener disconnects.

RendererOff
Error returned by the screen handle when the renderer disconnects.

Enums

ErrorKind
The kind of an error that may happen when executing a terminal operation.