Struct andiskaz::screen::Screen

pub struct Screen<'terminal> { /* fields omitted */ }

A locked screen terminal with exclusive access to it. By default, [ScreenData] locks and unlocks every operation. With this struct, a locked screen handle, one can execute many operations without locking and unlocking.

Implementations

impl<'terminal> Screen<'terminal>

pub fn size(&self) -> Coord2

Returns the current size of the screen.

pub fn valid_size(&self) -> bool

Returns whether the stored size is the actual size and the actual size is valid.

pub fn min_size(&self) -> Coord2

Returns the minimum size required for the screen.

pub fn set(&mut self, point: Coord2, tile: Tile)

Sets every attribute of a given Tile. This operation is buffered.

pub fn transform_colors<P>(&mut self, point: Coord2, transformer: P) where
P: PairTransformer,

Sets the colors of a given Tile. This operation is buffered.

pub fn update<F, T>(&mut self, point: Coord2, updater: F) -> T where
F: FnOnce(&mut Tile) -> T,

Applies an update function to a Tile. An update function gets access to a mutable reference of a Tile, updates it, and then the screen handles any changes made to it. This operation is buffered.

pub fn get(&self, point: Coord2) -> &Tile

Gets the attributes of a given Tile, regardless of being flushed to the screen yet or not.

pub fn clear(&mut self, background: Color)

Sets every Tile into a whitespace grapheme with the given color.

pub fn styled_text<P>(&mut self, tstring: &TermString, style: Style<P>) -> Coord where
P: PairTransformer,

Prints a grapheme-encoded text (a TermString) using some style options like ratio to the screen, color, margin and others. See Style.