Struct andiskaz::style::Style

pub struct Style<C = Color2>
where
    C: Updater,
{
    pub left_margin: Coord,
    pub right_margin: Coord,
    pub top_margin: Coord,
    pub bottom_margin: Coord,
    pub min_width: Coord,
    pub max_width: Coord,
    pub min_height: Coord,
    pub max_height: Coord,
    pub align_numer: Coord,
    pub align_denom: Coord,
    pub colors: C,
}

Alignment, margin and other settings for texts.

Fields
left_margin: Coord

Left margin.

right_margin: Coord

Right margin.

top_margin: Coord

Top margin.

bottom_margin: Coord

Bottom margin.

min_width: Coord

Minimum width.

max_width: Coord

Maximum width.

min_height: Coord

Minimum height.

max_height: Coord

Maximum height.

align_numer: Coord

Alignment align_numererator.

align_denom: Coord

Alignment align_denomominator.

colors: C

Foreground-background color pair.

Implementations

impl<C> Style<C>
where C: Updater,

pub fn with_colors(colors: C) -> Self

Creates a style with the given colors.

pub fn colors<D>(self, colors: D) -> Style<D>
where D: Updater,

Updates the style to the given color updater.

pub fn left_margin(self, left_margin: Coord) -> Self

Sets left margin.

pub fn right_margin(self, right_margin: Coord) -> Self

Sets right margin.

pub fn top_margin(self, top_margin: Coord) -> Self

Sets top margin.

pub fn bottom_margin(self, bottom_margin: Coord) -> Self

Sets bottom margin.

pub fn min_width(self, min_width: Coord) -> Self

Sets minimum width.

pub fn max_width(self, max_width: Coord) -> Self

Sets maximum width.

pub fn min_height(self, min_height: Coord) -> Self

Sets minimum height.

pub fn max_height(self, max_height: Coord) -> Self

Sets maximum height.

pub fn align(self, align_numer: Coord, align_denom: Coord) -> Self

Sets alignment. Numerator and align_denomominator are used such that line\[index\] * align_numer / align_denom == screen\[index\]

pub fn make_margin_below(&self) -> Vec2

Makes a coordinate pair that contains the margin dimensions that are “less”.

pub fn make_margin_above(&self) -> Vec2

Makes a coordinate pair that contains the margin dimensions that are “greater”.

pub fn make_min_size(&self) -> Vec2

Makes a coordinate pair that contains the minima sizes.

pub fn make_max_size(&self) -> Vec2

Makes a coordinate pair that contains the maxima sizes.

pub fn make_size(&self, screen_size: Vec2) -> Vec2

Makes a coordinate pair that contains the actual sizes.