Module andiskaz::ui::menu

This module exports items related to menus in the UI, such as more extensive menus, or just dialogs for OK/CANCEL.

Structs

Menu

A menu, with a list of options and potentially a cancel option.

Enums

DangerPromptOption

An item of a prompt about a dangerous action.

Traits

MenuOption

A trait representing a menu option.