Crate andiskaz

This crate provides basic utilities for writing applications with Terminal User Interface (TUI). It provides an event listener, and it provides a handle to a double buffered screen renderer.

Modules

color
This module provides colors that are usable with the terminal handle implemented by this library.

coord
Provides types related to the coordinate system.

error
This module exports error types used by the terminal handles.

event
This module defines input events of a terminal.

screen
This module defines screen related utilities.

string
This module provides a string type used to be printed to the terminal. It is called "TermString". You can use the macro tstring! to build a TermString as a shortcut for TermString::new_lossy.

style
This module provides styles for terminal text.

terminal
This crate exports a terminal terminal and its utilites.

Macros

tstring
Creates a TermString from string literal. Currently, invalid string get the unicode replacement character in their invalid characters. However, implementation may change to panic in those cases.

tstring_concat
Concatenates various TermString or TermString-like into a new TermString. It takes everything by reference, and it is possible to mix types.

Functions

emergency_restore
Best-effort function to restore the terminal in a panic.