Crate andiskaz

This crate provides basic utilities for writing applications with Terminal User Interface (TUI). It provides an event listener, and it provides a handle to a double buffered screen renderer.

Examples

This is an example for a simple Hello, World! application. For more examples, see examples folder.

use andiskaz::{
    color::Color2,
    emergency_restore,
    error::Error,
    event::Event,
    style::Style,
    terminal::Terminal,
    tstring,
};
use std::{panic, process::exit};

/// Asynchronous main of a tokio project.
#[tokio::main]
async fn main() {
    // Set panic hook so we can see the panic even if terminal was being
    // used in raw mode.
    panic::set_hook(Box::new(|info| {
        let _ = emergency_restore();
        eprintln!("{}", info);
    }));

    // Create a terminal with default settings and run it.
    let result = Terminal::run(term_main).await;
    // If error, print it out and exit with bad code.
    if let Ok(Err(error)) | Err(error) = result {
        eprintln!("{}", error);
        exit(-1);
    }
}

/// The terminal main function.
async fn term_main(mut term: Terminal) -> Result<(), Error> {
    // Allocate space for a string safe for being printed.
    let string = tstring!["Hello, World! Press any key..."];
    loop {
        // Lock the screen. The word "lock" is important here.
        let mut screen = term.screen.lock().await?;
        // Put our message.
        screen.styled_text(&string, Style::with_colors(Color2::default()));
        // Drop the screen handle. IMPORTANT. If we don't drop, it will
        // block things like the screen renderer.
        drop(screen);

        // Check if a key was pressed. We'll wait until an event happens. If
        // a key was pressed or an error happened, we exit. The only event
        // that makes this `if` fail is a resize event, so, we have to
        // re-print our message in the next iteration.
        if let Ok(Event::Key(_)) | Err(_) = term.events.listen().await {
            break;
        }
    }

    Ok(())
}

Modules

color
This module provides colors that are usable with the terminal handle implemented by this library.

coord
Provides types related to the coordinate system.

error
This module exports error types used by the terminal handles.

event
This module defines input events of a terminal.

screen
This module defines screen related utilities.

string
This module provides a string type used to be printed to the terminal. It is called "TermString". You can use the macro tstring! to build a TermString as a shortcut for TermString::new_lossy.

style
This module provides styles for terminal text.

terminal
This crate exports a terminal terminal and its utilites.

Macros

tstring
Creates a TermString from string literal. Currently, invalid string get the unicode replacement character in their invalid characters. However, implementation may change to panic in those cases.

tstring_concat
Concatenates various TermString or TermString-like into a new TermString. It takes everything by reference, and it is possible to mix types.

Functions

emergency_restore
Best-effort function to restore the terminal in a panic.