Module andiskaz::color

This module provides colors that are usable with the terminal handle implemented by this library.

Structs

AdaptBgToFg

Adapts the brightness of the background color to match the foreground color of a pair of colors (Color2). This means background is modified.

AdaptFgToBg

Adapts the brightness of the foreground color to match the background color of a pair of colors (Color2). This means foreground is modified.

BadBasicColor

Error returned when a BasicColor is attempted to be created with an invalid code.

BadCmyColor

Error returned when a CmyColor is attempted to be created with invalid channels.

BadGrayColor

Error returned when a GrayColor is attempted to be created with an invalid brightness.

Brightness

The brightness of a color.

CmyColor

A CMY (Cyan-Magenta-Yellow) color. The lower one of its component is, the more it subtracts.

Color2

A pair of colors (foreground and background).

Color8Bit

An 8-bit encoded color for the terminal.

ContrastBgWithFg

Contrasts the brightness of the background color against the foreground color of a pair of colors (Color2). This means background is modified.

ContrastFgWithBg

Contrasts the brightness of the foreground color against the background color of a pair of colors (Color2). This means foreground is modified.

GrayColor

A gray-scale color. Goes from white, to gray, to black.

RgbColor

An RGB color ((Red-Green-Blue)). This is an additive color model, where the value of a color channel is how much the channel is added to the color. All channels zeroed are black.

UpdateBg

Updates the background of a pair of colors (Color2) to the given color.

UpdateFg

Updates the foreground of a pair of colors (Color2) to the given color.

Enums

BasicColor

A basic color used by the terminal.

Color

A color usable in the terminal.

Color8BitKind

The kind of a color. enum representation of an 8-bit color.

Traits

ApproxBrightness

A trait for types that can approximate their brightness.

Updater

A function that updates a Color2.