Module andiskaz::color

This module provides colors that are usable with the terminal handle implemented by this library.

Modules

transform
This module defines color transformers. A color Transformer is simply a function over Colors. In fact, if you want to use a plain function as a Transformer, use FromFn. Analogously, there is a transformer over Color2s, the PairTransformer.

Structs

BadBasicColor
Error returned when a BasicColor is attempted to be created with an invalid code.

BadCmyColor
Error returned when a CmyColor is attempted to be created with invalid channels.

BadGrayColor
Error returned when a GrayColor is attempted to be created with an invalid brightness.

Brightness
The brightness of a color.

CmyColor
A CMY (Cyan-Magenta-Yellow) color. The lower one of its component is, the more it subtracts.

Color2
A pair of colors (foreground and background).

Color8
An 8-bit encoded color for the terminal.

GrayColor
A gray-scale color. Goes from white, to gray, to black.

RgbColor
An RGB color ((Red-Green-Blue)). This is an additive color model, where the value of a color channel is how much the channel is added to the color. All channels zeroed are black.

Enums

BasicColor
A basic color used by the terminal.

Color
A color usable in the terminal.

Color8Kind
The kind of a color. enum representation of an 8-bit color.

Traits

ApproxBrightness
A trait for types that can approximate their brightness.