Enum andiskaz::error::ErrorKind

#[non_exhaustive]pub enum ErrorKind {
    AlreadyRunning(AlreadyRunning),
    RendererOff(RendererOff),
    EventsOff(EventsOff),
    IO(Error),
    Fmt(Error),
    ParseInt(ParseIntError),
    Utf8(FromUtf8Error),
    Join(JoinError),
    Custom(Box<dyn ErrorTrait + Send + Sync>),
}

The kind of an error that may happen when executing a terminal operation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AlreadyRunning(AlreadyRunning)

Terminal services already running.

RendererOff(RendererOff)

Renderer disconnected.

EventsOff(EventsOff)

Event listener disconnected.

IO(Error)

This is an IO error.

Fmt(Error)

This is a formatting error.

ParseInt(ParseIntError)

This is an error from a bad integer parse attempt.

Utf8(FromUtf8Error)

This is an string from UTF-8 conversion error.

Join(JoinError)

This is an error from a bad join.

Custom(Box<dyn ErrorTrait + Send + Sync>)

A custom error, stored in a trait object.

Implementations

impl ErrorKind

pub fn as_dyn(&self) -> &(dyn ErrorTrait + Send + Sync + 'static)

Returns this error kind as a trait object.