Struct andiskaz::screen::Screen

pub struct Screen { /* fields omitted */ }

A handle to the screen. Used to write or read from the screen cells, as well to perform more advanced operations.

Implementations

impl Screen

pub fn min_size(&self) -> Coord2

Returns the minimum size required for the screen.

pub async fn lock<'screen>(
&'screen self
) -> Result<LockedScreen<'screen>, RendererOff>

Locks the screen handle. Most operations of the screen handle will lock the handle, perform the operation, and then unlock the handle. When one locks the screen manually, one can use the methods directly on the locked screen, without having to lock and unlock everytime.

pub async fn size<'_>(&'_ self) -> Result<Coord2, RendererOff>

Returns the current size of the screen.

pub async fn set<'_>(
&'_ mut self,
point: Coord2,
tile: Tile
) -> Result<(), RendererOff>

Sets every attribute of a given Tile. This operation is buffered.

pub async fn transform_colors<P, '_>(
&'_ mut self,
point: Coord2,
transformer: P
) -> Result<(), RendererOff> where
P: PairTransformer,

Sets the colors of a given Tile. This operation is buffered.

pub async fn update<F, T, '_>(
&'_ mut self,
point: Coord2,
updater: F
) -> Result<T, RendererOff> where
F: FnOnce(&mut Tile) -> T,

Applies an update function to a Tile. An update function gets access to a mutable reference of a Tile, updates it, and then the screen handles any changes made to it. This operation is buffered.

pub async fn get<'_>(&'_ self, point: Coord2) -> Result<Tile, RendererOff>

Gets the attributes of a given Tile, regardless of being flushed to the screen yet or not.

pub async fn clear<'_>(
&'_ mut self,
background: Color
) -> Result<(), RendererOff>

Sets every Tile into a whitespace grapheme with the given color.

pub async fn styled_text<P, '_, '_>(
&'_ mut self,
tstring: &'_ TermString,
style: Style<P>
) -> Result<Coord, RendererOff> where
P: PairTransformer,

Prints a grapheme-encoded text (a TermString) using some style options like ratio to the screen, color, margin and others. See Style.