Struct andiskaz::style::Style

pub struct Style<P> where
    P: PairTransformer,  {
    pub left_margin: Coord,
    pub right_margin: Coord,
    pub top_margin: Coord,
    pub bottom_margin: Coord,
    pub min_width: Coord,
    pub max_width: Coord,
    pub min_height: Coord,
    pub max_height: Coord,
    pub align_numer: Coord,
    pub align_denom: Coord,
    pub colors: P,
}

Alignment, margin and other settings for texts.

Fields
left_margin: Coord

Left margin.

right_margin: Coord

Right margin.

top_margin: Coord

Top margin.

bottom_margin: Coord

Bottom margin.

min_width: Coord

Minimum width.

max_width: Coord

Maximum width.

min_height: Coord

Minimum height.

max_height: Coord

Maximum height.

align_numer: Coord

Alignment align_numererator.

align_denom: Coord

Alignment align_denomominator.

colors: P

Foreground-background color pair.

Implementations

impl Style<Id>

pub fn new() -> Self

Default settings.

impl<P> Style<P> where
P: PairTransformer,

pub fn with_colors(colors: P) -> Self

Creates the style object from the color transformer.

pub fn left_margin(self, left_margin: Coord) -> Self

Sets left margin.

pub fn right_margin(self, right_margin: Coord) -> Self

Sets right margin.

pub fn top_margin(self, top_margin: Coord) -> Self

Sets top margin.

pub fn bottom_margin(self, bottom_margin: Coord) -> Self

Sets bottom margin.

pub fn min_width(self, min_width: Coord) -> Self

Sets minimum width.

pub fn max_width(self, max_width: Coord) -> Self

Sets maximum width.

pub fn min_height(self, min_height: Coord) -> Self

Sets minimum height.

pub fn max_height(self, max_height: Coord) -> Self

Sets maximum height.

pub fn align(self, align_numer: Coord, align_denom: Coord) -> Self

Sets alignment. Numerator and align_denomominator are used such that line\[index\] * align_numer / align_denom == screen\[index\]

pub fn colors<Q>(&self, colors: Q) -> Style<Q> where
Q: PairTransformer,

Sets foreground and background colors.

pub fn make_margin_below(&self) -> Coord2

Makes a coordinate pair that contains the margin dimensions that are "less".

pub fn make_margin_above(&self) -> Coord2

Makes a coordinate pair that contains the margin dimensions that are "greater".

pub fn make_min_size(&self) -> Coord2

Makes a coordinate pair that contains the minima sizes.

pub fn make_max_size(&self) -> Coord2

Makes a coordinate pair that contains the maxima sizes.

pub fn make_size(&self, screen_size: Coord2) -> Coord2

Makes a coordinate pair that contains the actual sizes.