Struct andiskaz::ui::input::InputDialog

pub struct InputDialog<F>where
    F: FnMut(char) -> bool,{

    pub filter: F,
    pub title: TermString,
    pub ok_label: TermString,
    pub cancel_label: TermString,
    pub buffer: TermString,
    pub max: Coord,
    pub title_colors: Color2,
    pub selected_colors: Color2,
    pub unselected_colors: Color2,
    pub cursor_colors: Color2,
    pub box_colors: Color2,
    pub bg: Color,
    pub title_y: Coord,
    pub pad_after_title: Coord,
    pub pad_after_box: Coord,
    pub pad_after_ok: Coord,

}

A dialog asking for user input, possibly filtered.

Fields
filter: F

Filter of the valid characters for the input dialog’s box.

title: TermString

The title of the input dialog.

ok_label: TermString

Label showed by the “OK” button (default “OK”).

cancel_label: TermString

Label showed by the “CANCEL” button (default “CANCEL”).

buffer: TermString

Initial buffer of the input dialog.

max: Coord

Maximum size of the input box.

title_colors: Color2

Colors of the title.

selected_colors: Color2

Selected option colors.

unselected_colors: Color2

Unselected option colors.

cursor_colors: Color2

Input box’s cursor colors.

box_colors: Color2

Input box colors.

bg: Color

Background of non-text areas.

title_y: Coord

Position of the title.

pad_after_title: Coord

Padding lines inserted after the title.

pad_after_box: Coord

Padding lines inserted after the box.

pad_after_ok: Coord

Padding lines inserted after the OK option.

Implementations

impl<F> InputDialog<F>where F: FnMut(char) -> bool,

pub fn new(title: TermString, buffer: TermString, max: Coord, filter: F) -> Self

Creates a new input dialog, with the given title, initial buffer, maximum input size, and filter function.

pub async fn select(&mut self, term: &mut Terminal) -> Result<TermString, Error>

Gets user input without possibility of canceling it, with cursor at 0.

pub async fn select_with_initial( &mut self, term: &mut Terminal, cursor: usize ) -> Result<TermString, Error>

Gets user input without possibility of canceling it, with given initial cursor.

pub async fn select_with_cancel( &mut self, term: &mut Terminal ) -> Result<Option<TermString>, Error>

Gets user input with the user possibly canceling it, with cursor at 0 and OK initially selected.

pub async fn select_cancel_initial( &mut self, term: &mut Terminal, cursor: usize, selected: InputDialogItem ) -> Result<Option<TermString>, Error>

Gets user input with the user possibly canceling it, with given initial cursor and given initially selected item.